# polzunov-feast-server/src/main/resources/openAPI.yaml
openapi: 3.0.3
info:
  title: Polzunov.Feast
  version: '0.1'
servers:
  - url: 'http://localhost:8080' #TODO change to real domain
security:
  - bearerAuth: [ ]
...
paths:
...
  /category:
    post:
      summary: Create new category
      operationId: addCategory
      tags:
        - category
        - admin
      requestBody:
        description: |
         Category to create. 
          - *Id can be every random integer because provided id will be substituted with generated one*
          - Category name must be unique. Uniqueness check is case-sensitive
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        200:
          description: Category created, *it will have generated id*
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/TokenUnauthorized'
        404:
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        409:
          description: Category with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update category by id
      operationId: updateCategoryById
      tags:
        - category
        - admin
      requestBody:
        description: Category object with new data. *Category name must be unique. Uniqueness check is case-sensitive*
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        200:
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          description: Admin or category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        409:
          description: Category with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: Get all categories
      operationId: getAllCategories
      security: [ ]
      tags:
        - category
      responses:
        200:
          description: Categories sent
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
  /category/{id}:
    parameters:
      - in: path
        name: id
        description: Category id
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get category by id
      operationId: getCategoryById
      security: [ ]
      tags:
        - category
      responses:
        200:
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete category by id
      operationId: deleteCategoryById
      tags:
        - category
        - admin
      responses:
        200:
          description: Category deleted
        401:
          $ref: '#/components/responses/TokenUnauthorized'
        404:
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        409:
          description: Cannot delete category, because there are events associated with it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  
  ############################################ route ###################################################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http #TODO change to https
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TokenUnauthorized:
      description: Token authorization failed, relogin required
    NotFound:
      description: Not found
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        objectValidationViolations:
          type: array
          items:
            $ref: '#/components/schemas/ObjectValidationViolation'
        fieldValidationViolations:
          type: array
          items:
            $ref: '#/components/schemas/FieldValidationViolation'
        httpAttributeValidationViolation:
          type: array
          items:
            $ref: '#/components/schemas/HttpAttributeValidationViolation'
    ObjectValidationViolation:
      type: object
      properties:
        message:
          type: string
    FieldValidationViolation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    HttpAttributeValidationViolation:
      type: object
      properties:
        attribute:
          type: string
        message:
          type: string
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...